requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.8
anyio==4.11.0
async-timeout==5.0.1
audioread==3.0.1
av==15.1.0
bitarray==3.7.1
bleak==2.1.1
blinker==1.9.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cloudpickle==3.1.1
colorama==0.4.6
Cython==3.1.4
dbus-fast==3.1.2
decorator==5.2.1
demucs==4.0.1
distro==1.9.0
dora_search==0.1.12
einops==0.8.1
exceptiongroup==1.3.0
fairseq==0.12.2
faiss-cpu==1.12.0
filelock==3.19.1
Flask==3.1.2
fsspec==2025.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hydra-core==1.0.7
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
julius==0.2.7
lameenc==1.8.1
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.45.0
lxml==6.0.2
MarkupSafe==3.0.2
mpmath==1.3.0
msgpack==1.1.1
networkx==3.4.2
numba==0.62.0
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
omegaconf==2.0.6
openai==1.109.1
openunmix==1.3.0
packaging==25.0
platformdirs==4.4.0
pooch==1.8.2
portalocker==3.2.0
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pygame==2.6.1
python-dotenv==1.1.1
pyworld==0.3.5
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
resampy==0.4.3
retrying==1.4.2
sacrebleu==2.5.1
scikit-learn==1.7.2
scipy==1.15.3
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
submitit==1.5.3
sympy==1.14.0
tabulate==0.9.0
threadpoolctl==3.6.0
torch==2.8.0
torchaudio==2.8.0
torchcrepe==0.0.24
tqdm==4.67.1
treetable==0.2.6
triton==3.4.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
